requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
gunicorn==20.1.0
pipenv==2022.11.11
psycopg2-binary==2.9.5


